Move test_components from test_robot_hardware to hardware_interface package (#288)
* Remove test_robot_hardware package and move example hardware components to test folder of hardware_interface.

* Remove package from ci and lint configuration

* Correct diff to CMakeLists.txt
