+++ .github/workflows/ci.yml
@@ -42,7 +42,6 @@ jobs:
             controller_manager_msgs
             hardware_interface
             ros2_control
-            test_robot_hardware
           vcs-repo-file-url: |
             https://raw.githubusercontent.com/${{ github.repository }}/${{ github.sha }}/ros2_control/ros2_control.repos
           colcon-mixin-name: coverage-gcc
