--- test_robot_hardware/CHANGELOG.rst
@@ -1,21 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package test_robot_hardware
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-0.1.3 (2021-01-21)
-------------------
-
-0.1.2 (2021-01-06)
-------------------
-
-0.1.1 (2020-12-23)
-------------------
-
-0.1.0 (2020-12-22)
-------------------
-* Remove components namespace and folders (`#258 <https://github.com/ros-controls/ros2_control/issues/258>`_)
-* Added starting of resources into CM and RM (`#240 <https://github.com/ros-controls/ros2_control/issues/240>`_)
-* Use resource manager (`#236 <https://github.com/ros-controls/ros2_control/issues/236>`_)
-* Replace RCUTILS\_ with RCLCPP\_ for logging (`#62 <https://github.com/ros-controls/ros2_control/issues/62>`_)
-* import test_robot_hardware
-* Contributors: Bence Magyar, Denis Štogl, Jafar Abdi, Jordan Palacios, Karsten Knese
