--- test_robot_hardware/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="2">
-  <name>test_robot_hardware</name>
-  <version>0.1.3</version>
-  <description>Description of test_robot_hardware</description>
-  <maintainer email="karsten@osrfoundation.org">Karsten Knese</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>hardware_interface</depend>
-  <depend>rclcpp</depend>
-
-  <test_depend>ament_cmake_gtest</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
