--- test_robot_hardware/test_robot_components.xml
@@ -1,20 +1,0 @@
-<library path="test_robot_hardware_components">
-
-  <class name="test_robot_hardware/TestSingleJointActuator" type="test_robot_hardware::TestSingleJointActuator" base_class_type="hardware_interface::ActuatorInterface">
-    <description>
-      Test actuator component for a single position joint
-    </description>
-  </class>
-
-  <class name="test_robot_hardware/TestForceTorqueSensor" type="test_robot_hardware::TestForceTorqueSensor" base_class_type="hardware_interface::SensorInterface">
-    <description>
-      Test sensor component emulating a force torque sensor
-    </description>
-  </class>
-
-  <class name="test_robot_hardware/TestTwoJointSystem" type="test_robot_hardware::TestTwoJointSystem" base_class_type="hardware_interface::SystemInterface">
-    <description>
-      Test system component for a two degree of freedom position joint robot
-    </description>
-  </class>
-</library>
